+++ central/policy/service/service_impl.go
@@ -314,10 +314,24 @@ func (s *serviceImpl) PatchPolicy(ctx context.Context, request *v1.PatchPolicyRe
 	return s.PutPolicy(ctx, policy)
 }
 
-// DeletePolicy deletes an policy from the system.
+// DeletePolicy deletes a policy from the system.
 func (s *serviceImpl) DeletePolicy(ctx context.Context, request *v1.ResourceByID) (*v1.Empty, error) {
 	if request.GetId() == "" {
 		return nil, errors.Wrap(errox.InvalidArgs, "A policy id must be specified to delete a Policy")
+	}
+
+	policy, exists, err := s.policies.GetPolicy(ctx, request.GetId())
+	if !exists {
+		// make repeated calls with the same policy ID idempotent
+		return &v1.Empty{}, nil
+	} else if err != nil {
+		// if any database error other than not exist occurs, bail early
+		return nil, errors.Wrap(err, "Database error while trying to delete policy")
+	}
+
+	// Note: default policies cannot be deleted, only disabled
+	if policy.IsDefault {
+		return nil, errors.Wrap(errox.InvalidArgs, "A default policy cannot be deleted. (You can disable a default policy, but not delete it.)")
 	}
 
 	if err := s.policies.RemovePolicy(ctx, request.GetId()); err != nil {